working on subscriptions function to get the proper ids needed for unsubscribing... currently not functioning
