+++ json-server.py
@@ -19,6 +19,7 @@ from views import (
     get_all_users,
     get_user_detail,
     create_subscription,
+    delete_subscription,
 )
 
 
@@ -54,12 +55,16 @@ class JSONServer(HandleRequests):
             if url["pk"] != 0:
                 response_body = view_follower_subscriptions(url["pk"])
                 return self.response(response_body, status.HTTP_200_SUCCESS.value)
-            return self.response("", status.HTTP_400_CLIENT_ERROR_BAD_REQUEST_DATA.value)
+            return self.response(
+                "", status.HTTP_400_CLIENT_ERROR_BAD_REQUEST_DATA.value
+            )
         elif url["requested_resource"] == "comments":
             if url["pk"] != 0:
                 response_body = view_post_comments(url["pk"])
                 return self.response(response_body, status.HTTP_200_SUCCESS.value)
-            return self.response("", status.HTTP_400_CLIENT_ERROR_BAD_REQUEST_DATA.value)
+            return self.response(
+                "", status.HTTP_400_CLIENT_ERROR_BAD_REQUEST_DATA.value
+            )
         elif url["requested_resource"] == "users":
             if url["pk"] != 0:
                 response_body = get_user_detail(url["pk"])
@@ -139,6 +144,13 @@ class JSONServer(HandleRequests):
         if url["requested_resource"] == "posts":
             if pk != 0:
                 successfully_deleted = delete_post(pk)
+                if successfully_deleted:
+                    return self.response(
+                        "", status.HTTP_204_SUCCESS_NO_RESPONSE_BODY.value
+                    )
+        elif url["requested_resource"] == "subscriptions":
+            if pk != 0:
+                successfully_deleted = delete_subscription(pk)
                 if successfully_deleted:
                     return self.response(
                         "", status.HTTP_204_SUCCESS_NO_RESPONSE_BODY.value
