+++ views/__init__.py
@@ -9,6 +9,9 @@ from .post import (
 )
 from .tag import view_all_tags, create_tag
 from .category import view_all_categories, create_category
-from .subscribers import view_follower_subscriptions
 from .comments import view_post_comments
-from .subscriptions import create_subscription
+from .subscriptions import (
+    create_subscription,
+    view_follower_subscriptions,
+    delete_subscription,
+)
